feat(content-picker): move anchor links back to content picker
