+++ src/admin/content-block/helpers/generators/anchor-links.ts
@@ -52,7 +52,12 @@ export const ANCHOR_LINKS_BLOCK_CONFIG = (position: number = 0): ContentBlockCon
 			},
 			buttonAction: {
 				label: i18n.t('admin/content-block/helpers/generators/anchor-links___link'),
-				editorType: ContentBlockEditor.AnchorLinkSelect,
+				editorType: ContentBlockEditor.ContentPicker,
+				editorProps: {
+					allowedTypes: ['ANCHOR_LINK'],
+					hideTargetSwitch: true,
+					hideTypeDropdown: true,
+				},
 			},
 		},
 	},
